Use $http in async example
