+++ examples/async/containers/app.js
@@ -1,5 +1,3 @@
-import * as AsyncActions from '../actions/async';
-
 export default function app() {
   return {
     restrict: 'E',
@@ -12,7 +10,7 @@ export default function app() {
 
 class AppController {
 
-  constructor($ngRedux, $scope) {
+  constructor($ngRedux, $scope, AsyncActions) {
     const unsubscribe = $ngRedux.connect(this.mapStateToThis, AsyncActions)((selectedState, actions) => {
       this.componentWillReceiveStateAndActions(selectedState, actions);
       Object.assign(this, selectedState, actions);
