+++ examples/async/package.json
@@ -16,7 +16,6 @@
   "homepage": "https://github.com/wbuchwalter/ngRedux#readme",
   "dependencies": {
     "angular": "^1.4.4",
-    "isomorphic-fetch": "^2.1.1",
     "ng-redux": "^3.0.0",
     "redux": "^3.0.0",
     "redux-logger": "^2.0.2",
