+++ examples/async/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import {
   SELECT_REDDIT, INVALIDATE_REDDIT,
   REQUEST_POSTS, RECEIVE_POSTS
-} from '../actions/async';
+} from '../constants/ActionTypes';
 
 function selectedReddit(state = 'angularjs', action) {
   switch (action.type) {
